Tweaks to dynamic memory sizing.

+++ implementations/meth1/sort/record_common.hh
@@ -17,12 +17,12 @@ class Record;
 class RecordS;
 
 namespace Rec {
-  constexpr size_t KEY_LEN = 10;
-  constexpr size_t VAL_LEN = 90;
-  constexpr size_t SIZE = KEY_LEN + VAL_LEN;
+  constexpr uint64_t KEY_LEN = 10;
+  constexpr uint64_t VAL_LEN = 90;
+  constexpr uint64_t SIZE = KEY_LEN + VAL_LEN;
 
-  constexpr size_t LOC_LEN = sizeof( uint64_t );
-  constexpr size_t SIZE_WITH_LOC = SIZE + LOC_LEN;
+  constexpr uint64_t LOC_LEN = sizeof( uint64_t );
+  constexpr uint64_t SIZE_WITH_LOC = SIZE + LOC_LEN;
 
   /* A Max or minimum record type. */
   enum limit_t { MAX, MIN };
